==> brain_RNAseq/scripts/11_eigengene-trait_relationships.R <==
#This script calculates consensus module eigengenes are correlates them with the external traits measured in this study,
#including colitis severity measures and circulating inflammatory markers, all of which can be found in traits.csv
