==> brain_RNAseq/scripts/12_eigengene-colon_module_relationships.R <==
#This script performs a similar analysis to that in 11, but instead of traits we're using distal colon consensus module eigengenes
#These colon modules are grouped into temporally discrete categories and come from 06 of the colon scripts
#The analysis from 11 reveals associations with measures that are all temporally similar (they all achieve an extreme at 7d/9d)
#This analysis was performed to evaluate other temporal profiles - for instance, a brain gene program may be regulated during the recovery phase after colitis
