==> brain_RNAseq/scripts/15_module_member_heatmaps.R <==
#This script ranks genes by their meta-module membership to their respective module
#Then, two plots are generated:
    #1: a bar chart showing the meta-module membership statistic of these genes
    #2: a heatmap showing the regional kME for each gene, to show similarities/differences between tissues
